requirements.txt
yara_python
pyinstaller
pyarmor
python-dotenv